+++ modules/elementary_functions/builtin/cpp/Gateway.cpp
@@ -108,6 +108,7 @@
 #include "triuBuiltin.hpp"
 #include "trilBuiltin.hpp"
 #include "signBuiltin.hpp"
+#include "ismemberBuiltin.hpp"
 //=============================================================================
 using namespace Nelson;
 //=============================================================================
@@ -292,7 +293,8 @@ static const nlsGateway gateway[] = {
         CPP_BUILTIN_WITH_EVALUATOR },
     { "sign", (void*)Nelson::ElementaryFunctionsGateway::signBuiltin, 1, 1,
         CPP_BUILTIN_WITH_EVALUATOR },
-
+    { "ismember", (void*)Nelson::ElementaryFunctionsGateway::ismemberBuiltin, 1, 2,
+        CPP_BUILTIN_WITH_EVALUATOR },
 };
 //=============================================================================
 NLSGATEWAYFUNC(gateway)