+++ modules/validators/src/cpp/ValidatorsInternal.cpp
@@ -62,6 +62,12 @@ setEvaluator(Evaluator* eval)
     _eval = eval;
 }
 //=============================================================================
+Evaluator*
+getEvaluator()
+{
+    return _eval;
+}
+//=============================================================================
 static std::wstring
 invalidPositionMessage(int argPosition)
 {